Ignore DHCP pools that are out of range
If the user changes the subnet of an interface then applies without adjusting the DHCP pool range/s to be in the new subnet, then an invalid dhcpd.conf is generated. The DHCP server complains about it and exits. If you have only 1 LAN with DHCP then it does not make a difference - you are not going to get DHCP whatever happens because there is no valid pool data. But if there are multiple LAN-style interfaces with DHCP then you can cause no DHCP-service on LAN when you are messing with settings on OPT1/OPT2...

I did this to myself last night, and even after rebooting get no DHCP for either my LAN or OPT1, just because the OPT1 pool settings were bad.

This change checks that the pool ranges are in the interface subnet. If they are not, then they are excluded from dhcpd.conf and a notice os filed. The user gets the flashing notice stuff on the webGUI to tell them what is wrong. And any other good DHCP interfaces+pools continue to work.