XSS/Security updates from upstream
The pfSense maintainers have provided a list of 30 or so security
changes related to the web GUI that have been made since the start of
the Bootstrap project.

I have reviewed each of the changes and found that only trivial changes
were required to a few files.

Ready for review