Provide separate checkboxes for tmp and var RAM disks
As suggested in forum: https://forum.pfsense.org/index.php?topic=96705.0
It is not hard to do - so here it is if someone thinks it is worth
having. e.g. on my SG-2440 full install that has plenty of free RAM I
could enable just /tmp memory disk and leave /var where it is on the
eMMC.
This is an integration and resubmit of PR #1765
Note: Leave this sitting here for now. @jimp made some suggestions in
that forum thread about maybe having a dropdown that has the 4
combinations all in a single dropdown list to choose from. It needs some
design thought before implementing. A change like this would not go in
2.2.* anyway, so there is plenty of time to decide what to do before
2.3.