Protect unset() with isset() in upgrade_config
Forum: https://forum.pfsense.org/index.php?topic=96827.0
Sometimes there can be unexpected stuff in a config. To avoid errors
like "PHP Fatal error: Cannot unset string offsets in
/etc/inc/upgrade_config.inc on line 291" it seems good to use
belts-and-braces checks with any unset() of associative array elements,
just in case the thing above is an empty string (instead of an array or
non-existent thing). An unnecessary error in processing the upgrade code
can mean missing other important upgrade stuff.
These are the places I could see in upgrade_config.inc where unset() is
not already protected by isset(). A lot of them are "that could not ever
go wrong", but IMHO it is worth protecting them all just to be sure.
This is a resubmit of PR #1773 after integrating with current master.