Redmine #5041 Improve reverse lookup zone name generation
Now 192.168.0.0/24 generates a reverse lookup of:
0.168.192.in-addr.arpa
Also previously something like 192.168.42.128/25 would have generated:
128.42.168.192.in-addr.arpa
which is not a valid zone.
Now that will generate:
42.168.192.in-addr.arpa
which is at least the zone concerned, even though it is inevitably
"bigger" than the forward zone.
For in-between subnet bit counts I have attempted to keep the current
behaviour:
172.16.4.0/22 becomes 4.16.172.in-addr.arpa
172.16.0.0/22 becomes 16.172.in-addr.arpa
172.16.0.0/16 becomes 16.172.in-addr.arpa
This is all very odd - because actually, for example, if it is a /22
then it needs to be expanded into 4 reverse lookup zones. But also what
can be put in dhcpd.conf maybe can be a super-set of the reverse lookup
name space and work fine?
I am not sure about all that, so best not to maybe break other working
configurations.