Reword text in etc/rc
The current text is not good English. A suggested re-wording is here.
There is a discussion that could be had about what actual action to take
here, and how good fsck is at automatically recovering from xyz. This PR
is not about that. I just thought I would suggest some better English
for this message.