Clarify the use of noshuntlaninterfaces and autoexcludelanaddress
The existing code here looks like it should be working correctly. Maybe it would help future maintainers if the field name used for gathering the data value in the UI and $POSTing it is changed to autoexcludelanaddress. Then the code reads a bit more easily where the logic value of noshuntlaninterfaces and autoexcludelanaddress is swapped back and forth.