Redmine #5200 be less aggressive about DHCP Pool Notice V2
This one will log_error() the DHCP pool message when it detects the inconsistency at the end of the setup wizard during reload all.
That way it can still be seen in the system log that this happened, and one day someone might chase down all the steps in the "reload all" process.
Compare this with https://github.com/pfsense/pfsense/pull/1935 and choose which way you would like to go.