Limit interface MTU to 1500 if no explicit MTU configured
Without this change, setting the MTU of a PPPoE interface with a
VLAN parent interface above 1492 will result in the MTU of other
VLANs on the same physical NIC that are not explicitly configured
to rise above 1500. It seems unwise - and a Principle of Least
Astonishment breach - for an MTU requiring Ethernet jumbo frames
to come about without explicit configuration.