Hide the delete row button when there is only 1 row remaining
When deleting the 2nd-last row this causes the Delete button for the last remaining row to be hidden. Having no delete button for a single-row alias means there is no temptation for the user to press it, and thus no need for them to get the message 'You may not delete the last row!'.
When adding the 2nd row the Delete buttons will appear again. In this case the user can delete any of the rows, leaving 1 or more remaining.