Explain which thing is blue
On the installed packages table, the Version column version number has a hyper-link to the package changelog. That appears blue all the time. It is when the package name in the "Name" column is blue that there is a newer version available.
If you can think of some other way to make this clear without writing so much, then go for it.