For reinstallall there is no pkg specified
Reinstall all packages is currently just displaying the current list of installed packages. It does not prompt to reinstall.
Fix this test so that it will not apply to reinstallall. For reinstallall there is no pkg specified in the $_GET. It is invoked from diag_backup.php like:
header("Location: pkg_mgr_install.php?mode=reinstallall");