Fix firewall log dynamic rule lookup master
for rows that are dynamically added as time goes by.
See https://github.com/pfsense/pfsense/pull/2014 for version for RELENG_2_2.

This is theoretically the fix. But due to other issues that mean dynamic updates to not happen at all in 2.3-ALPHA, I haven't been able to actually test it just yet.