diag_nanobsd do not show reference count 0
When the user has selected "Keep media mounted read/write at all times." then the value of $refcount will be zero even though is_writable() is true. That is a normal condition.
We only want to display the (Reference count n) text when the reference count is abnormal.
This is a "bug" in 2.2.* also, not a regression in 2.3-ALPHA. But it might as well just be changed in master 2.3. Not much point back-porting little crud like this to RELENG_2_2 nowadays.