unbound_statistics section should be written to unbound.conf
The stuff generated by unbound_statistics() goes into unbound.conf
It had got separated out and was going nowhere. This puts the behavior back the way it was yesterday.

Note: There is a separate issue about this because the UI has no place to actually select the statistics settings. There is some upgrade code that looks like it was converting statistics settings from the old Unbound package into custom commands in the integrated Unbound. So it looks like the unbound_statistics() function can/should be removed???