Improve diag_authentication.php redmine #5440
Move the code to select the default auth server so that it only happens the first time into the page. Now if the user selects a different auth server to test, and presses test, the results come back and the selected auth server is re-displayed (rather than the system default auth server being re-chosen every time).
Do not use address-of "&" when referring to the $config value - we do not ever intend to modify the config in here, so there seems no point in risking it by using a pointer.