Fix tooltip text for associated filter rules
1) Add a port forward with associated filter rule
2) Go to firewall_rules and hover over the icon on the left of the rule.
It says "traffic is ed".

The problem is that, for whatever reason, an associated rule does not have ['type'] specified, even in config.xml. It is that way in 2.2.5, so not a new thing or regression. I guess that the back-end rule generation works by assuming that a rule in config.xml that does not have 'type' defined should be a pass rule.

So we can't write code anywhere that assumes that every rule has a defined 'type'.

This text was different fixed stuff in 2.2.5, so there was no problem there.

In fixing this I also made the text explicit for each type. This will help with the gettext translation. The way it did it before, putting "ed" on the end, would not have worked for translation.