Make block bogon rule display consistent
The block private networks and anti-lockout rules display in the firewall rules display with the schedule column blank. But the block bogon networks has an asterisk. That looks odd. I noticed it just now when looking at firewall rules in 2.3 and looked back in 2.2.5 also - it is the same there.