Remove unused time-update-interval
There is no longer any code for a UI field in 2.3 master.
In RELENG_2_2 system.php the code for the field display is commented
out.
The commented out code has been removed in the bootstrap conversion.
This removes the various bits of leftover processing and validation code
and the setting from the default config.