Status IPsec inconsistent Logs tab
For some reason Status->IPsec has a "Logs" tab in addition to the
"Overview", "Leases", "SAD" and "SPD" tabs. The "Logs" tab takes you off
to the logs screens, and so when you click the tab all the other tabs
are gone and are replaced by all the "Logs" interface tabs.
That seems odd and inconsistent. Other Status displays do not have this
(e.g. Status->DHCP, Status->Load Balancer, Status->OpenVPN. With all
these, direct access to the associated log page is by the shortcut icon.
Status->IPsec also has the shortcut icon to go to the logs, and it works
fine.

So I am suggesting to remove the "Logs" tab here.