Remove special code for Status Captive Portal #5529
The Captive Portal Status page has been fixed up so it displays nice stuff when there are no Captive Portal Zones defined.
On my system this test did not work consistently anyway. When displaying the Dashboard, Status->Captive Portal was in the menu. When displaying Services->Captive Portal (with no zones defined) then Status->Captive Portal was not in the menu.
I think that don't care to waste time investigating exactly why that was, because actually this special test here is not really needed anyway.