Standardize In / Out queues help text
An "In" got changed to an "It" by https://github.com/pfsense/pfsense/commit/66066eda36fedf4df73bf03c6603ea76f9813327
This text already had most uses of the terms "In" and "Out" with capitals. That seems reasonable here so the text is clear about which of the words "in" and "out" are references to the "in" and "Out" queue rather than just ordinary uses of the English words.