Fix duplicate syslog messages - Redmine #5606 for RELENG_2_2
The generated syslog.conf needs to have the system.log selections combined into a single line in 2 places. The OP on redmine 5606 just showed 1 place in the diff there.
On my system, this gets rid of the duplicated message on login:
php-fpm[11074]: /index.php: Successful login for user 'my.name' from: 10.11.12.13

Note: for some reason old line 848 includes local4.none - but local4.none does not appear in old line 864. What is that about? Is that another different bug?
I did not mess with that, but I noticed that the OPs diff listing did also add in local4.none where it seems to be missing.