Internationalize more of config.console.inc
This code is a whole mix of some palces that use gettext() and other chunks of output that are hardcoded text between "EOD" markers.
I have got rid of the "EOD" blocks where there is a question asked - those needed to have a space at the end of the question. To meet code style guide it is easy if questions are all in their own echo statements and the space " " can be visibly coded in.