Use the _title var to make the widget title
Each widget has an include file that already has the required widget
title in a var like $dyn_dns_status_title $carp_status_title etc.
Use that var to set the displayed title of the widget, rather than the
stuff that was capitalizing the file name and then having a list of
special cases...
This meant the loop that includes the widget include files had to be
moved higher up in the code.