HTML Compliance - Collapsible Form Section HREF
Bad value for attribute href on element a: Illegal character in fragment: not a URL code point.

<a data-toggle="collapse" href="#<id> .panel-body">
Syntax of URL:Any URL. For example: /hello, #canvas, or http://example.org/. Characters should be represented in NFC and spaces should be escaped as %20.

Fix is to set an id on the target.