Update system_advanced_firewall.php
A few little code style bits
Make the State Timeouts section actually appear (add the section to the form)
The State Timeouts section was trying to do groups of fields, which spread themselves horizontally. That did not fit across a reasonable screen nicely, specially for the 6 different TCP timeouts. So the groups thing is removed and each field is its own input in the section, all vertically below each other.
Fixed various cut-and-paste bugs in the variable naming of the timeouts code.
Added logic to be able to generate the key name (like tcpfinwaittimeout) from the screen displayed text (like "FIN Wait"). Previously the "FIN Wait" case would not work with the simple strtolower($name)

This works now. If someone wants to mess with grouping the TCP, UDP, ICMP and Other in some way, then great.