HTML Compliance - Services Status Widget
Attribute height not allowed on element select at this point.

The for attribute of the label element must refer to a non-hidden form control.
'<label for="inputPassword3" class="col-sm-3 control-label">Hidden services</label>'
Can't find an "inputPassword3" form element this binds too.  Maybe it's a placeholder for future feature.  Leaving it for now.