dyn_dns_status widget avoid var naming issue
This use of var $i was conflicting and interfering with a use of it in
index.php
Use a much more unique name for now to avoid future risk.
Maybe a proper final solution for var naming across dashboard and widget
is needed - but that is a design issue to be discussed.