system_usermanager remove unused code
1) Line 448 "Remove this certificate association? (Certificate will not be deleted)" there was the text "Delete" being displayed and not looking good, in addition to the trash bin icon. Just the trash bin icon seems to be all that is needed.
2) Old line 757 there was a weird-looking "else;" - that seemed to be effectively an empty "else" clause hanging off the end of the "if" block above it. In that case it does nothing and I have removed it.
3) Remove the extra indent of a whole lot of code below old line 757 that actually was not in any "else" clause. This lines it up the way it actually works.
4) Old line 758
$section = new Form_Section('User Certificates');
That made a new $section and chunks of code below it added stuff to the section and then it was thrown away - there was nothing that actually added the section tp the form. That code is similar to what was already above that uses build_cert_table() to make a new Form_StaticText().
So I deleted the code from old line 758 that was functionally useless.

I think it all still works :)

Please review this carefully and make sure I have done good things. The code seemed a bit of a mess to me - maybe it was, or maybe I have no idea!