WebGUI Left Column Labels Hyper
Touch Screen Usability Annoyance
https://forum.pfsense.org/index.php?topic=104135.0

Clicking the labels in the left column selects/activates the first item.  This is the perfectly natural place for scrolling with thumb.

No big problem really for input fields, but check boxes and buttons are inadvertently selected/toggled resulting in unintended configuration changes.

Would be real nice if the labels in the left column were not "active".

This commit provides an option in System - General Setup page to enable/disable this "feature".

There is probably a better way, but this appears to be functional.  For now at least it seems to work for me anyway.  Just have to eliminate (null) the for attribute on the group labels so it doesn't have a target.