Avoid triggering initial wizard before reboot
If you use the the webGUI to reset to factory defaults, then while the
existing system is shutting down you navigate off to the dashboard, the
initial setup wizard will start. The trigger_initial_wizard flag file is
deleted, and so after the reboot the initial setup wizard will not be
triggered.
This change fixes that sequence of events by using a different flag file
across the reboot, and switching it back to the usual flag file name
during the subsequent boot up.