Feature #5729 Add Description to DHCP Leases list RELENG_2_2
The start and end columns are n/a for static leases, and it is static leases for which we (might) have a description. So we could reuse the Start and End columns to show the description. That avoids any issue of using more screen width and how to fit all the columns.
I tried it at home on my 2.2.6 system and actually it is useful for me - the descriptions can be a lot more descriptive than the host name that I have put in the static mappings, so it is easier to know what is what on the status DHCP leases display with the description shown.
This is just an example based on RELENG_2_2 - but it might be the way to go?