HTML Compliance - Firewall / NAT / Outbound / Edit
Bad value [0-9, a-z, A-Z and . for attribute pattern on element input: Unterminated character class
<input class="form-control" name="source" id="source" type="text" pattern="[0-9, a-z, A-Z and .">
<input class="form-control" name="destination" id="destination" type="text" value="" pattern="[0-9, a-z, A-Z and .">

Element ul not allowed as child of element span in this context. (Suppressing further errors from this subtree.)
Contexts in which element ul may be used:
 Where flow content is expected.
Content model for element span:
 Phrasing content.

Element br not allowed as child of element ul in this context. (Suppressing further errors from this subtree.)
Contexts in which element br may be used:
 Where phrasing content is expected.
Content model for element ul:
 Zero or more li and script-supporting elements.