Config history info - make it an info box
Currently this display as an orange box. It seems to me that it is ordinary info and should be an info box (displaying in blue).
Note: with the current behavior the text is always displayed, after changing it to an info box the "i" icon has to be clicked to show the text.
Do we sometimes(like in this example) want the info text to be automagically displayed on initial page load?